--- rustfmt.toml
@@ -1,2 +1,0 @@
-max_width = 70
-tab_spaces = 2
