+++ src/inet_utils.rs
@@ -3,46 +3,42 @@ use reqwest::{Client, Error};
 use std::time::Instant;
 
 pub async fn is_connected() -> bool {
-  let response = reqwest::get("https://www.google.com").await;
+    let response = reqwest::get("https://www.google.com").await;
 
-  match response {
-    Ok(res) => res.status().is_success(),
-    Err(_) => false,
-  }
+    match response {
+        Ok(res) => res.status().is_success(),
+        Err(_) => false,
+    }
 }
 
-pub async fn measure_download_speed(
-  url: String,
-) -> Result<f64, Error> {
-  let client = Client::new();
-  let start_time = Instant::now();
-  let response = client.get(url).send().await?.error_for_status()?;
+pub async fn measure_download_speed(url: String) -> Result<f64, Error> {
+    let client = Client::new();
+    let start_time = Instant::now();
+    let response = client.get(url).send().await?.error_for_status()?;
 
-  let content_length = response.content_length().unwrap_or(0);
-  let duration = start_time.elapsed();
+    let content_length = response.content_length().unwrap_or(0);
+    let duration = start_time.elapsed();
 
-  let speed_mbps = (content_length as f64 / 1024.0 / 1024.0)
-    / duration.as_secs_f64();
-  Ok(speed_mbps.clone())
+    let speed_mbps = (content_length as f64 / 1024.0 / 1024.0) / duration.as_secs_f64();
+    Ok(speed_mbps.clone())
 }
 
 pub async fn measure_upload_speed(url: &str) -> Result<f64, Error> {
-  let args = scic::Args::parse();
+    let args = args::Args::parse();
 
-  let client = Client::new();
-  let data = vec![0u8; args.size * 1024 * 1024];
+    let client = Client::new();
+    let data = vec![0u8; args.size * 1024 * 1024];
 
-  let start_time = Instant::now();
-  let _response = client
-    .post(url)
-    .body(data.clone())
-    .send()
-    .await?
-    .error_for_status()?;
+    let start_time = Instant::now();
+    let _response = client
+        .post(url)
+        .body(data.clone())
+        .send()
+        .await?
+        .error_for_status()?;
 
-  let duration = start_time.elapsed();
-  let speed_mbps =
-    (data.len() as f64 / 1024.0 / 1024.0) / duration.as_secs_f64();
+    let duration = start_time.elapsed();
+    let speed_mbps = (data.len() as f64 / 1024.0 / 1024.0) / duration.as_secs_f64();
 
-  Ok(speed_mbps)
+    Ok(speed_mbps)
 }
